V 1.01 -Added a cheap form of sanitization on the TimeZone User Input to prevent the clock timer from frying itself from Unhandled Exceptions. Those are handled now and a error is displayed to the user if the timezone text they entered is incorrect.
-Extended the "box" of the clock so it has room for the - AM or - PM 
text. Previously if you were using the AM/PM mode and the clock reached
Anything above 1 number on the hour dial, the AM/PM text would be 
replaced with 3 dots as there was not enough room to display the 
remaning text of the clock, that has been fixed.

New Features:

Fucking nothing, it's dead in here.